refactor: 모코코 스케줄러를 이벤트 핸들러로 분리
